Created new file for object for Skill, adjusted farming.js to no include functions in skillObject.js
Added a new file to handle the function and data for skill objects, and rewrote the function to be handle as object and store values. This will allow for reducing reduncy when adding other skills other then farming. Also linked the new .js file to the farming.html and gave type of module.