Add Setters and Getters to skill class
Added 4 function which set and get the values within the class. All, except the setter for the skill field, just call the required function to recheck the actually values using the function for getting there respective values. When the skill is changed the all of the other setters are called to refresh the values stored within the class. Added error checking to make sure the setter for the class is actually a valid skill. Throw the error "Invalid Skill Entered"